+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/list/Messages.kt
@@ -29,10 +29,6 @@ import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -230,11 +226,11 @@ internal fun BoxScope.DefaultMessagesHelperContent(
 
     val coroutineScope = rememberCoroutineScope()
 
-    val firstVisibleItemIndex by remember { derivedStateOf { lazyListState.firstVisibleItemIndex } }
+    val firstVisibleItemIndex = lazyListState.firstVisibleItemIndex
 
     val focusedItemIndex = messages.indexOfFirst { it is MessageItemState && it.focusState is MessageFocused }
 
-    val offset by messagesLazyListState.focusedMessageOffset.collectAsState(0)
+    val offset = messagesLazyListState.focusedMessageOffset
 
     LaunchedEffect(newMessageState, focusedItemIndex, offset) {