+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/state/channel/internal/ChannelMutableState.kt
@@ -76,6 +76,7 @@ internal class ChannelMutableState(
     private val _insideSearch = MutableStateFlow(false)
     private val _loadingOlderMessages = MutableStateFlow(false)
     private val _loadingNewerMessages = MutableStateFlow(false)
+    private val _lastSentMessageDate = MutableStateFlow<Date?>(null)
 
     /** Channel config data. */
     private val _channelConfig: MutableStateFlow<Config> = MutableStateFlow(Config())
@@ -171,7 +172,7 @@ internal class ChannelMutableState(
     override val membersCount: StateFlow<Int> = _membersCount
 
     override val channelData: StateFlow<ChannelData> =
-        _channelData.filterNotNull().combine(latestUsers) { channelData, users ->
+        combine(_channelData.filterNotNull(), latestUsers) { channelData, users ->
             if (users.containsKey(channelData.createdBy.id)) {
                 channelData.copy(createdBy = users[channelData.createdBy.id] ?: channelData.createdBy)
             } else {
@@ -183,8 +184,6 @@ internal class ChannelMutableState(
             ChannelData(
                 type = channelType,
                 channelId = channelId,
-                insideSearch = _insideSearch.value,
-                cachedLatestMessages = cachedLatestMessages.value.values.toList()
             )
         )
 
@@ -201,21 +200,27 @@ internal class ChannelMutableState(
 
     override val insideSearch: StateFlow<Boolean> = _insideSearch
 
+    override val lastSentMessageDate: StateFlow<Date?> = _lastSentMessageDate
+
     override fun toChannel(): Channel {
         // recreate a channel object from the various observables.
         val channelData = channelData.value
 
-        val messages = if (_insideSearch.value) cachedLatestMessages.value.values.toList() else sortedMessages.value
+        val messages = sortedMessages.value
+        val cachedMessages = cachedLatestMessages.value.values.toList()
         val members = members.value
         val watchers = watchers.value
         val reads = _rawReads.value.values.toList()
         val watcherCount = _watcherCount.value
+        val insideSearch = _insideSearch.value
 
-        val channel = channelData.toChannel(messages, members, reads, watchers, watcherCount)
+        val channel = channelData
+            .toChannel(messages, cachedMessages, members, reads, watchers, watcherCount, insideSearch)
         channel.config = _channelConfig.value
         channel.unreadCount = unreadCount.value
-        channel.lastMessageAt = messages.lastOrNull()?.let { it.createdAt ?: it.createdLocallyAt }
         channel.hidden = _hidden.value
+        channel.isInsideSearch = _insideSearch.value
+        channel.cachedLatestMessages = cachedLatestMessages.value.values.toList()
 
         return channel
     }
@@ -319,6 +324,15 @@ internal class ChannelMutableState(
         }
 
         _insideSearch.value = isInsideSearch
+    }
+
+    /**
+     * Sets the date of the last message sent by the current user.
+     *
+     * @param lastSentMessageDate The date of the last message.
+     */
+    fun setLastSentMessageDate(lastSentMessageDate: Date?) {
+        _lastSentMessageDate.value = lastSentMessageDate
     }
 
     /**
@@ -470,7 +484,7 @@ internal class ChannelMutableState(
     }
 
     private fun cacheLatestMessages() {
-        cachedLatestMessages.value = sortedVisibleMessages.value.associateBy(Message::id)
+        cachedLatestMessages.value = sortedMessages.value.associateBy(Message::id)
     }
 
     /**